refactor: remove tenant impersonate functionality and consolidate services
