+++ packages/apps/frontera/src/infra/repositories/common/common.datum.ts
@@ -1,4 +1,9 @@
 import { GetSlackChannelsQuery } from './queries/getSlackChannels.generated';
+import { TenantImpersonateListQuery } from '../common/queries/impersonateList.generated.ts';
 
 export type SlackChannelDatum =
   GetSlackChannelsQuery['slackChannelsWithBot'][0];
+
+export type TenantImpersonateDatum = NonNullable<
+  TenantImpersonateListQuery['tenant_impersonateList'][0]
+>;
