+++ packages/apps/frontera/src/infra/repositories/common/mutations/switchCurrentWorkspace.generated.ts
