+++ packages/apps/frontera/src/infra/repositories/common/queries/impersonateList.generated.ts
