--- packages/apps/frontera/src/infra/repositories/tenant/index.ts
@@ -1,1 +1,0 @@
-export * from './tenant-impersonate.repository.ts';
