--- packages/apps/frontera/src/infra/repositories/tenant/tenant-impersonate.datum.ts
@@ -1,5 +1,0 @@
-import { TenantImpersonateListQuery } from './queries/impersonateList.generated.ts';
-
-export type TenantImpersonateDatum = NonNullable<
-  TenantImpersonateListQuery['tenant_impersonateList'][0]
->;
